refactor(github_client): split transport from resource helpers (Codex #3)
Move Label dataclass + list/create/update/delete_label helpers out of
github_client.py into a new github_api/labels.py module. github_client.py
now contains only the generic gh CLI subprocess transport (run_gh,
run_gh_api) and the GhError classification hierarchy.

Rationale (Codex PR #10 architectural finding #3): the transport layer
and resource-specific helpers have different reasons to change.
github_client grew label-shaped during Phase 5 and would grow further
when Phase 6+ adds issues/PRs/workflows. Separating them keeps each
file focused on one responsibility.

No behavior change. All 112 tests pass after import path updates.

Layout:
  github_client.py         — transport + error hierarchy
  github_api/labels.py     — Label dataclass + CRUD helpers
  github_api/__init__.py   — empty package marker

Test layout mirrors source:
  tests/unit/github_client/  — transport tests only
  tests/unit/github_api/     — resource helper tests
