+++ tests/unit/github_client/test_github_client.py
@@ -1,8 +1,13 @@
-"""Tests for gh_manage.github_client with subprocess.run mocked."""
+"""Tests for gh_manage.github_client — transport + error hierarchy only.
+
+Resource-specific label tests (list_labels, create_label, update_label,
+delete_label) moved to tests/unit/github_api/test_labels.py during the
+Phase 5 checkpoint refactor. This file tests only the generic transport
+layer: run_gh, run_gh_api, and the GhError classification.
+"""
 
 from __future__ import annotations
 
-import json
 from subprocess import CompletedProcess
 
 import pytest
@@ -16,13 +21,8 @@ from gh_manage.github_client import (
     GhNotInstalledError,
     GhPermissionError,
     GhRateLimitError,
-    Label,
-    create_label,
-    delete_label,
-    list_labels,
     run_gh,
     run_gh_api,
-    update_label,
 )
 
 
@@ -42,164 +42,7 @@ def _mock_gh_failure(mocker: MockerFixture, stderr: str, returncode: int = 1):
     )
 
 
-# Happy path — list_labels.
-# NOTE: Phase 5 list_labels uses `gh api --paginate --jq '.[]'` which emits
-# one JSON object per line (NDJSON). Tests mock stdout with newline-separated
-# JSON objects, not a single JSON array.
-def test_list_labels_parses_ndjson_response(mocker: MockerFixture) -> None:
-    ndjson = (
-        json.dumps({"name": "bug", "color": "d73a4a", "description": "Buggy"})
-        + "\n"
-        + json.dumps({"name": "feat", "color": "a2eeef", "description": None})
-        + "\n"
-    )
-    _mock_gh_success(mocker, ndjson)
-    result = list_labels("yakkuro/gh-manage")
-    assert result == [
-        Label(name="bug", color="d73a4a", description="Buggy"),
-        Label(name="feat", color="a2eeef", description=""),
-    ]
-
-
-def test_list_labels_uses_paginate_and_jq_flags(mocker: MockerFixture) -> None:
-    """list_labels must use `--paginate --jq '.[]'` to produce NDJSON.
-
-    Plain `--paginate` emits multiple JSON documents concatenated which
-    json.loads cannot parse for repos with >100 labels. `--jq '.[]'`
-    makes gh emit one JSON object per line — safe to parse line-by-line.
-    Regression for Codex HIGH finding on PR #8.
-    """
-    mock_run = _mock_gh_success(mocker, "")
-    list_labels("yakkuro/gh-manage")
-    args = mock_run.call_args.args[0]
-    assert "--paginate" in args
-    assert "--jq" in args
-    # The jq filter should flatten the page array to individual items
-    jq_idx = args.index("--jq")
-    assert args[jq_idx + 1] == ".[]"
-
-
-def test_list_labels_handles_empty_response(mocker: MockerFixture) -> None:
-    """Empty stdout → empty list."""
-    _mock_gh_success(mocker, "")
-    result = list_labels("yakkuro/gh-manage")
-    assert result == []
-
-
-def test_list_labels_handles_multi_page_response(mocker: MockerFixture) -> None:
-    """Regression test for Codex HIGH finding: a multi-page response via
-    --paginate --jq '.[]' outputs one JSON object per line. Repos with
-    >100 labels (or any multi-page result) must parse correctly."""
-    ndjson_lines = [
-        json.dumps({"name": f"label{i}", "color": "000000", "description": ""})
-        for i in range(250)  # simulates 3 pages × ~100 labels
-    ]
-    _mock_gh_success(mocker, "\n".join(ndjson_lines) + "\n")
-    result = list_labels("yakkuro/big-repo")
-    assert len(result) == 250
-    assert result[0].name == "label0"
-    assert result[249].name == "label249"
-
-
-def test_list_labels_raises_gh_api_error_on_malformed_ndjson_line(
-    mocker: MockerFixture,
-) -> None:
-    """If gh api produces a malformed line (API format change, truncated
-    response), list_labels must raise GhAPIError, not propagate
-    json.JSONDecodeError as a raw traceback."""
-    ndjson = (
-        json.dumps({"name": "bug", "color": "d73a4a", "description": ""})
-        + "\n"
-        + "{this is not valid json\n"
-    )
-    _mock_gh_success(mocker, ndjson)
-    with pytest.raises(GhAPIError, match="Failed to parse label entry"):
-        list_labels("yakkuro/gh-manage")
-
-
-# Normalization
-def test_list_labels_normalizes_color_to_lowercase(mocker: MockerFixture) -> None:
-    _mock_gh_success(
-        mocker,
-        json.dumps({"name": "bug", "color": "D73A4A", "description": "x"}) + "\n",
-    )
-    result = list_labels("yakkuro/gh-manage")
-    assert result[0].color == "d73a4a"
-
-
-def test_list_labels_converts_null_description_to_empty_string(
-    mocker: MockerFixture,
-) -> None:
-    _mock_gh_success(
-        mocker,
-        json.dumps({"name": "bug", "color": "d73a4a", "description": None}) + "\n",
-    )
-    result = list_labels("yakkuro/gh-manage")
-    assert result[0].description == ""
-
-
-# Happy path — create_label
-def test_create_label_sends_correct_body(mocker: MockerFixture) -> None:
-    mock_run = _mock_gh_success(mocker, "")
-    create_label(
-        "yakkuro/gh-manage",
-        Label(name="chore", color="e1e7eb", description="housekeeping"),
-    )
-    args = mock_run.call_args.args[0]
-    assert "api" in args
-    assert "repos/yakkuro/gh-manage/labels" in args
-    assert "-X" in args
-    assert "POST" in args
-    assert "name=chore" in args
-    assert "color=e1e7eb" in args
-    assert "description=housekeeping" in args
-
-
-# Happy path — update_label with rename
-def test_update_label_with_rename_includes_new_name(mocker: MockerFixture) -> None:
-    mock_run = _mock_gh_success(mocker, "")
-    update_label(
-        "yakkuro/gh-manage",
-        current_name="bug",
-        new_label=Label(name="fix", color="d73a4a", description="Bug fix"),
-    )
-    args = mock_run.call_args.args[0]
-    assert "repos/yakkuro/gh-manage/labels/bug" in args
-    assert "-X" in args
-    assert "PATCH" in args
-    assert "new_name=fix" in args
-    assert "color=d73a4a" in args
-    assert "description=Bug fix" in args
-
-
-# Happy path — update_label without rename
-def test_update_label_without_rename_omits_new_name(mocker: MockerFixture) -> None:
-    mock_run = _mock_gh_success(mocker, "")
-    update_label(
-        "yakkuro/gh-manage",
-        current_name="fix",
-        new_label=Label(name="fix", color="d73a4a", description="Updated desc"),
-    )
-    args = mock_run.call_args.args[0]
-    assert "repos/yakkuro/gh-manage/labels/fix" in args
-    assert "-X" in args
-    assert "PATCH" in args
-    assert not any("new_name=" in a for a in args)
-    assert "color=d73a4a" in args
-    assert "description=Updated desc" in args
-
-
-# Happy path — delete_label
-def test_delete_label_calls_correct_endpoint(mocker: MockerFixture) -> None:
-    mock_run = _mock_gh_success(mocker, "")
-    delete_label("yakkuro/gh-manage", "bug")
-    args = mock_run.call_args.args[0]
-    assert "repos/yakkuro/gh-manage/labels/bug" in args
-    assert "-X" in args
-    assert "DELETE" in args
-
-
-# Error classification — parametrized
+# Error classification — parametrized over all 6 stderr patterns
 @pytest.mark.parametrize(
     ("stderr", "expected_exc"),
     [
@@ -243,15 +86,14 @@ def test_gh_auth_error_mentions_gh_auth_login(mocker: MockerFixture) -> None:
         run_gh_api("repos/foo/bar/labels")
 
 
-# Regression: silent-failure-hunter review findings
+# Regression: silent-failure-hunter review findings (Phase 5)
 def test_run_gh_api_malformed_json_raises_gh_api_error(
     mocker: MockerFixture,
 ) -> None:
-    """Regression test for silent-failure-hunter MEDIUM finding:
-    json.loads on malformed stdout must be wrapped into GhAPIError so the
-    user sees an actionable message instead of a raw JSONDecodeError
-    traceback. This can happen in practice with truncated responses or
-    GitHub API format changes."""
+    """json.loads on malformed stdout must be wrapped into GhAPIError so
+    the user sees an actionable message instead of a raw JSONDecodeError
+    traceback. Can happen in practice with truncated responses or GitHub
+    API format changes."""
     _mock_gh_success(mocker, "{this is not valid json")
     with pytest.raises(GhAPIError, match="invalid JSON"):
         run_gh_api("repos/foo/bar/labels")
@@ -260,8 +102,7 @@ def test_run_gh_api_malformed_json_raises_gh_api_error(
 def test_run_gh_non_zero_exit_propagates_classified_error(
     mocker: MockerFixture,
 ) -> None:
-    """Regression test for silent-failure-hunter HIGH finding:
-    run_gh (the lower-level function) must always raise a GhError subclass
+    """run_gh (the low-level function) must always raise a GhError subclass
     on non-zero exit — never return silently. Tests exercise run_gh_api
     which wraps run_gh; this test directly verifies run_gh itself so a
     future refactor can't silently regress the non-zero path."""
